function_app/requirements.txt
azure-functions
azure-identity
azure-storage-blob
pandas
requests
